mysite/requirements.txt
annotated-types==0.7.0
asgiref==3.9.1
attrs==25.3.0
certifi==2025.8.3
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
Django==5.2.5
django-taggit==5.0.1
frictionless==5.18.1
humanize==4.13.0
idna==3.10
isodate==0.7.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
Markdown==3.6
markdown-it-py==4.0.0
marko==2.2.0
MarkupSafe==3.0.2
mdurl==0.1.2
petl==1.7.17
psycopg==3.1.18
psycopg-binary==3.1.18
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.4
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
rfc3986==2.0.0
rich==14.1.0
rpds-py==0.27.1
shellingham==1.5.4
simpleeval==1.0.3
six==1.17.0
sqlparse==0.5.3
tabulate==0.9.0
text-unidecode==1.3
typer==0.17.3
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
validators==0.35.0
